:white_check_mark: :poop: fix container views test suite
Somehow, the generic test scenario for `bulk edit` does not work if the member variable `bulk_edit_data` does not contain as **first** key a field that is not set in the seed data.

Thus, adding `"hostname": "h"` to replace the previously removed `"state": "running"` fixes the test, but we are not sure why.